Moved models out from /database to sit with associated routes and controllers. Improved authentication component hierarchy to differentiate between service auth and csb auth
