+++ src/components/authentication/csb-access-token/csb-token-authentication.js
@@ -3,9 +3,9 @@ import CustomStrategy from 'passport-custom';
 
 import uuid from 'node-uuid';
 import moment from 'moment';
-import accessToken from '../../database/models/access-token';
-import log from '../../log';
-import * as userController from '../user/user-controller';
+import accessToken from './csb-token-model';
+import log from '../../../log';
+import * as userController from '../../user/user-controller';
 
 export function generateTokenForUser(userId) {
   return new Promise((resolve, reject) => {
