+++ src/components/user/user-controller.js
@@ -1,4 +1,4 @@
-import User from '../../database/models/user';
+import User from './user-model';
 import log from '../../log';
 
 const PUBLIC_API_ATTRIBUTES = [
