+++ src/components/user/user-routes.js
@@ -1,6 +1,6 @@
 import express from 'express';
 import * as users from './user-controller';
-import { csbTokenAuthenticatedOnly } from '../authentication/csb-token-authentication';
+import { csbTokenAuthenticatedOnly } from '../authentication/csb-access-token/csb-token-authentication';
 
 const router = express.Router(); // eslint-disable-line new-cap
